// src/project-v0/mobile/auth/auth.dto.ts
export class saveUserDto {
    phone_number: String;
    username: String;
    password: String;
    site_id: String;
    site_name: String;
    firstname: String;
    lastname: String;
    emp_id: Number;
    emp_position: String;
    image: String;
}

export class loginDto {
    username: String;
    password: String;
}